moved the leaderboard files to the correctly names leaderboard package
