+++ src/main/java/it/unibo/exam/model/leaderboard/LeaderboardEntry.java
@@ -1,4 +1,4 @@
-package it.unibo.exam.model;
+package it.unibo.exam.model.leaderboard;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
